Move execution-time-only option lookup to service
Relocates the logic for finding the task that declares an
execution-time-only option from `ExecutionTimeOnlyOptionsValidationException`
to `ExecutionTimeOnlyOptionsManifestService`. This centralizes
responsibilities for execution-time-only options within the manifest
service, improving code organization and separation of concerns.
